blog_publisher/blog_publisher/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose
python-multipart